LPS-45659 remove threadlocal usages from trashcan methods and deprecate class

+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLAppHelperLocalServiceImpl.java
@@ -67,7 +67,6 @@ import com.liferay.portlet.documentlibrary.model.DLSyncEvent;
 import com.liferay.portlet.documentlibrary.service.DLConfig;
 import com.liferay.portlet.documentlibrary.service.base.DLAppHelperLocalServiceBaseImpl;
 import com.liferay.portlet.documentlibrary.social.DLActivityKeys;
-import com.liferay.portlet.documentlibrary.util.DLAppHelperThreadLocal;
 import com.liferay.portlet.documentlibrary.util.DLProcessorRegistryUtil;
 import com.liferay.portlet.documentlibrary.util.DLUtil;
 import com.liferay.portlet.documentlibrary.util.comparator.FileVersionVersionComparator;
@@ -76,8 +75,11 @@ import com.liferay.portlet.trash.model.TrashEntry;
 import com.liferay.portlet.trash.model.TrashVersion;
 import com.liferay.portlet.trash.util.TrashUtil;
 
+import javax.portlet.PortletPreferences;
+import javax.portlet.PortletRequest;
+import javax.portlet.PortletURL;
+import javax.servlet.http.HttpServletRequest;
 import java.io.Serializable;
-
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
@@ -85,12 +87,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.Callable;
-
-import javax.portlet.PortletPreferences;
-import javax.portlet.PortletRequest;
-import javax.portlet.PortletURL;
-
-import javax.servlet.http.HttpServletRequest;
 
 /**
  * Provides the local service helper for the document library application.
@@ -654,6 +650,17 @@ public class DLAppHelperLocalServiceImpl
 			ServiceContext serviceContext)
 		throws PortalException, SystemException {
 
+		return moveFileEntryFromTrash(
+			userId, fileEntry, newFolderId, serviceContext,
+			DLConfig.getRestrictedDLConfig());
+	}
+
+	@Override
+	public FileEntry moveFileEntryFromTrash(
+			long userId, FileEntry fileEntry, long newFolderId,
+			ServiceContext serviceContext, DLConfig dlConfig)
+		throws PortalException, SystemException {
+
 		boolean hasLock = dlFileEntryLocalService.hasFileEntryLock(
 			userId, fileEntry.getFileEntryId());
 
@@ -664,7 +671,7 @@ public class DLAppHelperLocalServiceImpl
 
 		try {
 			return doMoveFileEntryFromTrash(
-				userId, fileEntry, newFolderId, serviceContext);
+				userId, fileEntry, newFolderId, serviceContext, dlConfig);
 		}
 		finally {
 			if (!hasLock) {
@@ -687,6 +694,25 @@ public class DLAppHelperLocalServiceImpl
 	public FileEntry moveFileEntryToTrash(long userId, FileEntry fileEntry)
 		throws PortalException, SystemException {
 
+		return moveFileEntryToTrash(
+			userId, fileEntry, DLConfig.getLiberalDLConfig());
+	}
+
+	/**
+	 * Moves the file entry to the recycle bin.
+	 *
+	 * @param  userId the primary key of the user moving the file entry
+	 * @param  fileEntry the file entry to be moved
+	 * @param  dlConfig the DL configuration object
+	 * @return the moved file entry
+	 * @throws PortalException if a user with the primary key could not be found
+	 * @throws SystemException if a system exception occurred
+	 */
+	@Override
+	public FileEntry moveFileEntryToTrash(
+			long userId, FileEntry fileEntry, DLConfig dlConfig)
+		throws PortalException, SystemException {
+
 		boolean hasLock = dlFileEntryLocalService.hasFileEntryLock(
 			userId, fileEntry.getFileEntryId());
 
@@ -696,7 +722,7 @@ public class DLAppHelperLocalServiceImpl
 		}
 
 		try {
-			return doMoveFileEntryToTrash(userId, fileEntry);
+			return doMoveFileEntryToTrash(userId, fileEntry, dlConfig);
 		}
 		finally {
 			if (!hasLock) {
@@ -1084,6 +1110,15 @@ public class DLAppHelperLocalServiceImpl
 	public void restoreFileEntryFromTrash(long userId, FileEntry fileEntry)
 		throws PortalException, SystemException {
 
+		restoreFileEntryFromTrash(
+			userId, fileEntry, DLConfig.getLiberalDLConfig());
+	}
+
+	@Override
+	public void restoreFileEntryFromTrash(
+			long userId, FileEntry fileEntry, DLConfig dlConfig)
+		throws PortalException, SystemException {
+
 		// File entry
 
 		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
@@ -1100,18 +1135,19 @@ public class DLAppHelperLocalServiceImpl
 
 		dlFileEntryLocalService.updateStatus(
 			userId, fileVersion.getFileVersionId(), trashEntry.getStatus(),
-			new HashMap<String, Serializable>(), new ServiceContext());
+			new HashMap<String, Serializable>(), dlConfig,
+			new ServiceContext());
 
-		if (DLAppHelperThreadLocal.isEnabled()) {
-
-			// File rank
-
+		if (dlConfig.isFileRanksEnabled()) {
 			dlFileRankLocalService.enableFileRanks(fileEntry.getFileEntryId());
+		}
 
-			// File shortcut
-
+		if (dlConfig.isFileShortcutsEnabled()) {
 			dlFileShortcutLocalService.enableFileShortcuts(
 				fileEntry.getFileEntryId());
+		}
+
+		if (dlConfig.isDLSyncEventEnabled()) {
 
 			// Sync
 
@@ -1121,36 +1157,34 @@ public class DLAppHelperLocalServiceImpl
 
 		// Trash
 
-		List<TrashVersion> trashVersions = trashVersionLocalService.getVersions(
-			trashEntry.getEntryId());
+		if (dlConfig.isTrashEnabled()) {
+			List<TrashVersion> trashVersions =
+				trashVersionLocalService.getVersions(trashEntry.getEntryId());
 
-		for (TrashVersion trashVersion : trashVersions) {
-			DLFileVersion trashDLFileVersion =
-				dlFileVersionPersistence.findByPrimaryKey(
-					trashVersion.getClassPK());
+			for (TrashVersion trashVersion : trashVersions) {
+				DLFileVersion trashDLFileVersion =
+					dlFileVersionPersistence.findByPrimaryKey(
+						trashVersion.getClassPK());
 
-			trashDLFileVersion.setStatus(trashVersion.getStatus());
+				trashDLFileVersion.setStatus(trashVersion.getStatus());
 
-			dlFileVersionPersistence.update(trashDLFileVersion);
+				dlFileVersionPersistence.update(trashDLFileVersion);
+			}
+
+			trashEntryLocalService.deleteEntry(trashEntry.getEntryId());
 		}
 
-		trashEntryLocalService.deleteEntry(trashEntry.getEntryId());
+		if (dlConfig.isSocialActivityEnabled()) {
+			JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
 
-		if (!DLAppHelperThreadLocal.isEnabled()) {
-			return;
+			extraDataJSONObject.put("title", fileEntry.getTitle());
+
+			socialActivityLocalService.addActivity(
+				userId, fileEntry.getGroupId(),
+				DLFileEntryConstants.getClassName(), fileEntry.getFileEntryId(),
+				SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
+				extraDataJSONObject.toString(), 0);
 		}
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
-
-		extraDataJSONObject.put("title", fileEntry.getTitle());
-
-		socialActivityLocalService.addActivity(
-			userId, fileEntry.getGroupId(), DLFileEntryConstants.getClassName(),
-			fileEntry.getFileEntryId(),
-			SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
 	}
 
 	@Override
@@ -1508,13 +1542,9 @@ public class DLAppHelperLocalServiceImpl
 	public void updateStatus(
 			long userId, FileEntry fileEntry, FileVersion latestFileVersion,
 			int oldStatus, int newStatus,
-			Map<String, Serializable> workflowContext,
+			Map<String, Serializable> workflowContext, DLConfig dlConfig,
 			ServiceContext serviceContext)
 		throws PortalException, SystemException {
-
-		if (!DLAppHelperThreadLocal.isEnabled()) {
-			return;
-		}
 
 		if (newStatus == WorkflowConstants.STATUS_APPROVED) {
 
@@ -1522,7 +1552,9 @@ public class DLAppHelperLocalServiceImpl
 
 			String latestFileVersionVersion = latestFileVersion.getVersion();
 
-			if (latestFileVersionVersion.equals(fileEntry.getVersion())) {
+			if (dlConfig.isAssetEnabled() &&
+				latestFileVersionVersion.equals(fileEntry.getVersion())) {
+
 				if (!latestFileVersionVersion.equals(
 						DLFileEntryConstants.VERSION_DEFAULT)) {
 
@@ -1576,9 +1608,9 @@ public class DLAppHelperLocalServiceImpl
 
 			// Sync
 
-			String event = (String)workflowContext.get("event");
+			String event = (String) workflowContext.get("event");
 
-			if (Validator.isNotNull(event)) {
+			if (dlConfig.isDLSyncEventEnabled()&& Validator.isNotNull(event)) {
 				registerDLSyncEventCallback(event, fileEntry);
 			}
 
@@ -1587,34 +1619,44 @@ public class DLAppHelperLocalServiceImpl
 
 				// Social
 
-				Date activityCreateDate = latestFileVersion.getModifiedDate();
-				int activityType = DLActivityKeys.UPDATE_FILE_ENTRY;
+				if (dlConfig.isSocialActivityEnabled()) {
+					Date activityCreateDate =
+						latestFileVersion.getModifiedDate();
 
-				if (event.equals(DLSyncConstants.EVENT_ADD)) {
-					activityCreateDate = latestFileVersion.getCreateDate();
-					activityType = DLActivityKeys.ADD_FILE_ENTRY;
+					int activityType = DLActivityKeys.UPDATE_FILE_ENTRY;
+
+					if (event.equals(DLSyncConstants.EVENT_ADD)) {
+						activityCreateDate = latestFileVersion.getCreateDate();
+						activityType = DLActivityKeys.ADD_FILE_ENTRY;
+					}
+
+					JSONObject extraDataJSONObject =
+						JSONFactoryUtil.createJSONObject();
+
+					extraDataJSONObject.put("title", fileEntry.getTitle());
+
+					socialActivityLocalService.addUniqueActivity(
+						latestFileVersion.getStatusByUserId(),
+						fileEntry.getGroupId(), activityCreateDate,
+						DLFileEntryConstants.getClassName(),
+						fileEntry.getFileEntryId(), activityType,
+						extraDataJSONObject.toString(), 0);
 				}
-
-				JSONObject extraDataJSONObject =
-					JSONFactoryUtil.createJSONObject();
-
-				extraDataJSONObject.put("title", fileEntry.getTitle());
-
-				socialActivityLocalService.addUniqueActivity(
-					latestFileVersion.getStatusByUserId(),
-					fileEntry.getGroupId(), activityCreateDate,
-					DLFileEntryConstants.getClassName(),
-					fileEntry.getFileEntryId(), activityType,
-					extraDataJSONObject.toString(), 0);
 
 				// Subscriptions
 
-				notifySubscribers(latestFileVersion, serviceContext);
+				if (dlConfig.isSubscriptionEnabled()) {
+					notifySubscribers(latestFileVersion, serviceContext);
+				}
 			}
 		}
 		else {
 
 			// Asset
+
+			if (!dlConfig.isAssetEnabled()) {
+				return;
+			}
 
 			boolean visible = false;
 
@@ -1635,9 +1677,22 @@ public class DLAppHelperLocalServiceImpl
 		}
 	}
 
+	@Override
+	public void updateStatus(
+			long userId, FileEntry fileEntry, FileVersion latestFileVersion,
+			int oldStatus, int newStatus,
+			Map<String, Serializable> workflowContext,
+			ServiceContext serviceContext)
+		throws PortalException, SystemException {
+
+		updateStatus(
+			userId, fileEntry, latestFileVersion, oldStatus, newStatus,
+			workflowContext, DLConfig.getLiberalDLConfig(), serviceContext);
+	}
+
 	protected FileEntry doMoveFileEntryFromTrash(
 			long userId, FileEntry fileEntry, long newFolderId,
-			ServiceContext serviceContext)
+			ServiceContext serviceContext, DLConfig dlConfig)
 		throws PortalException, SystemException {
 
 		// File entry
@@ -1651,7 +1706,7 @@ public class DLAppHelperLocalServiceImpl
 				userId, fileEntry.getFileEntryId(), newFolderId,
 				serviceContext);
 
-			if (DLAppHelperThreadLocal.isEnabled()) {
+			if (dlConfig.isFileRanksEnabled()) {
 				dlFileRankLocalService.enableFileRanks(
 					fileEntry.getFileEntryId());
 			}
@@ -1682,7 +1737,7 @@ public class DLAppHelperLocalServiceImpl
 
 		dlFileEntryLocalService.updateStatus(
 			userId, fileVersion.getFileVersionId(), oldStatus,
-			new HashMap<String, Serializable>(), serviceContext);
+			new HashMap<String, Serializable>(), dlConfig, serviceContext);
 
 		// File versions
 
@@ -1711,14 +1766,11 @@ public class DLAppHelperLocalServiceImpl
 			}
 		}
 
-		if (DLAppHelperThreadLocal.isEnabled()) {
-
-			// File rank
-
+		if (dlConfig.isFileRanksEnabled()) {
 			dlFileRankLocalService.enableFileRanks(fileEntry.getFileEntryId());
+		}
 
-			// File shortcut
-
+		if (dlConfig.isFileShortcutsEnabled()) {
 			dlFileShortcutLocalService.enableFileShortcuts(
 				fileEntry.getFileEntryId());
 		}
@@ -1728,26 +1780,28 @@ public class DLAppHelperLocalServiceImpl
 		fileEntry = dlAppService.moveFileEntry(
 			fileEntry.getFileEntryId(), newFolderId, serviceContext);
 
-		// Sync
+		if (dlConfig.isDLSyncEventEnabled()) {
+			registerDLSyncEventCallback(
+				DLSyncConstants.EVENT_RESTORE, fileEntry);
+		}
 
-		registerDLSyncEventCallback(DLSyncConstants.EVENT_RESTORE, fileEntry);
+		if (dlConfig.isSocialActivityEnabled()) {
+			JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
 
-		// Social
+			extraDataJSONObject.put("title", fileEntry.getTitle());
 
-		JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
-
-		extraDataJSONObject.put("title", fileEntry.getTitle());
-
-		socialActivityLocalService.addActivity(
-			userId, fileEntry.getGroupId(), DLFileEntryConstants.getClassName(),
-			fileEntry.getFileEntryId(),
-			SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
-			extraDataJSONObject.toString(), 0);
+			socialActivityLocalService.addActivity(
+				userId, fileEntry.getGroupId(),
+				DLFileEntryConstants.getClassName(), fileEntry.getFileEntryId(),
+				SocialActivityConstants.TYPE_RESTORE_FROM_TRASH,
+				extraDataJSONObject.toString(), 0);
+		}
 
 		return fileEntry;
 	}
 
-	protected FileEntry doMoveFileEntryToTrash(long userId, FileEntry fileEntry)
+	protected FileEntry doMoveFileEntryToTrash(
+			long userId, FileEntry fileEntry, DLConfig dlConfig)
 		throws PortalException, SystemException {
 
 		// File versions
@@ -1766,84 +1820,82 @@ public class DLAppHelperLocalServiceImpl
 		dlFileEntryLocalService.updateStatus(
 			userId, fileVersion.getFileVersionId(),
 			WorkflowConstants.STATUS_IN_TRASH,
-			new HashMap<String, Serializable>(), new ServiceContext());
+			new HashMap<String, Serializable>(), dlConfig,
+			new ServiceContext());
 
-		if (DLAppHelperThreadLocal.isEnabled()) {
-
-			// File shortcut
-
+		if (dlConfig.isFileShortcutsEnabled()) {
 			dlFileShortcutLocalService.disableFileShortcuts(
 				fileEntry.getFileEntryId());
+		}
 
-			// File rank
-
+		if (dlConfig.isFileRanksEnabled()) {
 			dlFileRankLocalService.disableFileRanks(fileEntry.getFileEntryId());
+		}
+
+		if (dlConfig.isDLSyncEventEnabled()) {
 
 			// Sync
 
 			registerDLSyncEventCallback(DLSyncConstants.EVENT_TRASH, fileEntry);
 		}
 
-		// Trash
+		if (dlConfig.isTrashEnabled()) {
+			List<ObjectValuePair<Long, Integer>> dlFileVersionStatusOVPs =
+				new ArrayList<ObjectValuePair<Long, Integer>>();
 
-		int oldDLFileVersionStatus = WorkflowConstants.STATUS_ANY;
+			DLFileVersion oldDLFileVersion =
+				(DLFileVersion)fileVersion.getModel();
 
-		List<ObjectValuePair<Long, Integer>> dlFileVersionStatusOVPs =
-			new ArrayList<ObjectValuePair<Long, Integer>>();
+			int oldDLFileVersionStatus = oldDLFileVersion.getStatus();
 
-		DLFileVersion oldDLFileVersion = (DLFileVersion)fileVersion.getModel();
+			if ((dlFileVersions != null) && !dlFileVersions.isEmpty()) {
+				dlFileVersionStatusOVPs = getDlFileVersionStatuses(
+					dlFileVersions);
+			}
 
-		oldDLFileVersionStatus = oldDLFileVersion.getStatus();
+			for (DLFileVersion curDLFileVersion : dlFileVersions) {
+				curDLFileVersion.setStatus(WorkflowConstants.STATUS_IN_TRASH);
 
-		if ((dlFileVersions != null) && !dlFileVersions.isEmpty()) {
-			dlFileVersionStatusOVPs = getDlFileVersionStatuses(dlFileVersions);
+				dlFileVersionPersistence.update(curDLFileVersion);
+			}
+
+			DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
+
+			UnicodeProperties typeSettingsProperties = new UnicodeProperties();
+
+			typeSettingsProperties.put("title", dlFileEntry.getTitle());
+
+			TrashEntry trashEntry = trashEntryLocalService.addTrashEntry(
+				userId, dlFileEntry.getGroupId(),
+				DLFileEntryConstants.getClassName(),
+				dlFileEntry.getFileEntryId(), dlFileEntry.getUuid(),
+				dlFileEntry.getClassName(), oldDLFileVersionStatus,
+				dlFileVersionStatusOVPs, typeSettingsProperties);
+
+			String trashTitle = TrashUtil.getTrashTitle(
+				trashEntry.getEntryId());
+
+			dlFileEntry.setTitle(trashTitle);
+
+			dlFileEntryPersistence.update(dlFileEntry);
 		}
 
-		for (DLFileVersion curDLFileVersion : dlFileVersions) {
-			curDLFileVersion.setStatus(WorkflowConstants.STATUS_IN_TRASH);
+		if (dlConfig.isSocialActivityEnabled()) {
+			JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
 
-			dlFileVersionPersistence.update(curDLFileVersion);
+			extraDataJSONObject.put(
+				"title", TrashUtil.getOriginalTitle(fileEntry.getTitle()));
+
+			socialActivityLocalService.addActivity(
+				userId, fileEntry.getGroupId(),
+				DLFileEntryConstants.getClassName(), fileEntry.getFileEntryId(),
+				SocialActivityConstants.TYPE_MOVE_TO_TRASH,
+				extraDataJSONObject.toString(), 0);
 		}
 
-		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
+		if (dlConfig.isWorkflowEnabled() &&
+			(oldStatus == WorkflowConstants.STATUS_PENDING)) {
 
-		UnicodeProperties typeSettingsProperties = new UnicodeProperties();
-
-		typeSettingsProperties.put("title", dlFileEntry.getTitle());
-
-		TrashEntry trashEntry = trashEntryLocalService.addTrashEntry(
-			userId, dlFileEntry.getGroupId(),
-			DLFileEntryConstants.getClassName(), dlFileEntry.getFileEntryId(),
-			dlFileEntry.getUuid(), dlFileEntry.getClassName(),
-			oldDLFileVersionStatus, dlFileVersionStatusOVPs,
-			typeSettingsProperties);
-
-		String trashTitle = TrashUtil.getTrashTitle(trashEntry.getEntryId());
-
-		dlFileEntry.setTitle(trashTitle);
-
-		dlFileEntryPersistence.update(dlFileEntry);
-
-		if (!DLAppHelperThreadLocal.isEnabled()) {
-			return fileEntry;
-		}
-
-		// Social
-
-		JSONObject extraDataJSONObject = JSONFactoryUtil.createJSONObject();
-
-		extraDataJSONObject.put(
-			"title", TrashUtil.getOriginalTitle(fileEntry.getTitle()));
-
-		socialActivityLocalService.addActivity(
-			userId, fileEntry.getGroupId(), DLFileEntryConstants.getClassName(),
-			fileEntry.getFileEntryId(),
-			SocialActivityConstants.TYPE_MOVE_TO_TRASH,
-			extraDataJSONObject.toString(), 0);
-
-		// Workflow
-
-		if (oldStatus == WorkflowConstants.STATUS_PENDING) {
 			workflowInstanceLinkLocalService.deleteWorkflowInstanceLink(
 				fileVersion.getCompanyId(), fileVersion.getGroupId(),
 				DLFileEntryConstants.getClassName(),
